% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{az_conn_settings}
\alias{az_conn_settings}
\title{Get Azure settings from a DuckDB connection}
\usage{
az_conn_settings(conn = az_conn())
}
\arguments{
\item{conn}{A DuckDB connection. Defaults to \code{\link[=az_conn]{az_conn()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{name}, \code{value}, \code{description}.
}
\description{
Queries \code{duckdb_settings()} and returns all entries whose name contains
\code{"azure"}.
}
\examples{
conn <- DBI::dbConnect(duckdb::duckdb())
az_conn_settings(conn)
DBI::dbDisconnect(conn, shutdown = TRUE)
}
