% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{conn_default}
\alias{conn_default}
\title{Default DuckDB connection}
\usage{
conn_default()
}
\value{
A DuckDB connection.
}
\description{
Returns the DuckDB default connection (thin wrapper around
\code{\link[duckdb:default_conn]{duckdb::default_conn()}}).
}
\keyword{internal}
