% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{conn_open}
\alias{conn_open}
\title{Create a DuckDB connection}
\usage{
conn_open(..., drv = conn_driver(), timezone_out = "", array = "matrix")
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{drv}{A DuckDB driver. Defaults to \code{\link[=conn_driver]{conn_driver()}}.}

\item{timezone_out}{Character scalar. Timezone for \code{TIMESTAMPTZ} output.
Defaults to \code{""} (UTC).}

\item{array}{Character scalar. How to represent DuckDB arrays.
Defaults to \code{"matrix"}.}
}
\value{
A \code{duckdb_connection} object.
}
\description{
Thin wrapper around \code{\link[DBI:dbConnect]{DBI::dbConnect()}} with quak-friendly defaults.
}
\keyword{internal}
