% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{ext_cache_path}
\alias{ext_cache_path}
\title{Default DuckDB extension cache directory}
\usage{
ext_cache_path()
}
\value{
Character scalar. The resolved cache path.
}
\description{
Resolution order: in-memory value (\code{opts$set("cache_dir", ...)}) ->
env var \code{QUAK_CACHE_DIR} -> OS-appropriate user cache directory via
\code{\link[tools:R_user_dir]{tools::R_user_dir()}}.
}
\examples{
ext_cache_path()
}
