% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{ext_dir}
\alias{ext_dir}
\title{Find the DuckDB extension folder}
\usage{
ext_dir(conn = conn_default())
}
\arguments{
\item{conn}{A DuckDB connection. Defaults to \code{\link[=conn_default]{conn_default()}}.}
}
\value{
Character scalar. Path to the extension directory.
}
\description{
Returns the path where DuckDB stores installed extension files.
This is determined by the \code{extension_directory} setting.
}
\examples{
conn <- DBI::dbConnect(duckdb::duckdb())
ext_dir(conn)
DBI::dbDisconnect(conn, shutdown = TRUE)
}
