% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{ext_is_installed}
\alias{ext_is_installed}
\title{Check whether a DuckDB extension is installed}
\usage{
ext_is_installed(name, conn = conn_default())
}
\arguments{
\item{name}{Character scalar. Extension name.}

\item{conn}{A DuckDB connection. Defaults to \code{\link[=conn_default]{conn_default()}}.}
}
\value{
Logical scalar.
}
\description{
Check whether a DuckDB extension is installed
}
\examples{
conn <- DBI::dbConnect(duckdb::duckdb())
ext_is_installed("httpfs", conn = conn)
DBI::dbDisconnect(conn, shutdown = TRUE)
}
