% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{ext_list_available}
\alias{ext_list_available}
\title{List all DuckDB core extensions}
\usage{
ext_list_available(conn = conn_default())
}
\arguments{
\item{conn}{A DuckDB connection. Defaults to \code{\link[=conn_default]{conn_default()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{name}, \code{version}, \code{description}.
}
\description{
Returns the full catalog of extensions maintained by the DuckDB core team,
regardless of whether they are installed.
}
\examples{
conn <- DBI::dbConnect(duckdb::duckdb())
ext_list_available(conn)
DBI::dbDisconnect(conn, shutdown = TRUE)
}
