% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{ext_uninstall}
\alias{ext_uninstall}
\title{Uninstall a DuckDB extension}
\usage{
ext_uninstall(
  name,
  purge_cache = FALSE,
  cache = ext_cache(),
  conn = conn_default()
)
}
\arguments{
\item{name}{Character scalar. Extension name.}

\item{purge_cache}{Logical. If \code{TRUE}, also removes the file from \code{cache}.}

\item{cache}{An \code{ext_cache} object. Only used when \code{purge_cache = TRUE}.}

\item{conn}{A DuckDB connection. Defaults to \code{\link[=conn_default]{conn_default()}}.}
}
\value{
Invisibly returns \code{conn}.
}
\description{
Removes the extension file from DuckDB's \code{extension_directory}. Optionally
also purges the corresponding entry from the local cache.
}
\examples{
\dontrun{
# Requires a connection with the extension already installed.
conn <- DBI::dbConnect(duckdb::duckdb())
ext_uninstall("httpfs", conn = conn)
DBI::dbDisconnect(conn, shutdown = TRUE)
}
}
