% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quak-package.R
\docType{package}
\name{quak-package}
\alias{quak}
\alias{quak-package}
\title{quak: Query 'Azure Data Lake Storage Gen2' with 'DuckDB'}
\description{
Provides convenience utilities for using 'DuckDB' directly over datasets stored in 'Azure Data Lake Storage Gen2' (ADLS Gen2, 'abfss://'). Opens connections configured for Azure-backed 'Delta Lake' and 'Parquet' data, registers Azure credentials as 'DuckDB' secrets, and supports optional repository mirrors for restricted networks. Integrates well with 'DBI' for SQL workflows and with 'dplyr' and 'dbplyr' for lazy table queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/quak/}
  \item \url{https://github.com/pedrobtz/quak}
  \item Report bugs at \url{https://github.com/pedrobtz/quak/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

}
\keyword{internal}
